feat: restart connection
