+++ qb/qcount/count.go
@@ -3,6 +3,7 @@ package qcount
 import (
 	"github.com/gocql/gocql"
 
+	"github.com/Drafteame/cassandra-builder/qb/models"
 	"github.com/Drafteame/cassandra-builder/qb/query"
 )
 
@@ -10,6 +11,7 @@ import (
 
 type Client interface {
 	Session() *gocql.Session
+	Config() models.Config
 	Debug() bool
 	Restart() error
 	PrintFn() query.DebugPrint
